+++ frontend/src/scenes/Workspace/routines/routines.ts
@@ -1,4 +1,4 @@
 import { createRoutine } from 'redux-saga-routines';
 
 export const addWorkspaceRoutine = createRoutine('ADD_WORKSPACE');
-export const userChannelRoutine = createRoutine('USER_SELECT_CHANNEL');
+export const selectChatRoutine = createRoutine('SELECT_CHAT');
